agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "strands-agents>=1.0.0",
]

[dependency-groups]
dev = [
    "pandas>=2.3.1",
    "pytest>=8.4.1",
]
